+++ main_test.go
@@ -209,3 +209,4 @@ func TestSchemaVersion(t *testing.T) {
 		}
 	}
 }
+